Update the "fee_percentage" to u16 value in "cosmwasm_tokenwrapper"
Update the constant "WRAP_FEE_CALC_DENOMINATOR" to "10_000_u16"

+++ contracts/tokenwrapper/src/contract.rs
@@ -458,7 +458,7 @@ fn update_config(
     governor: Option<String>,
     is_native_allowed: Option<bool>,
     wrapping_limit: Option<Uint128>,
-    fee_percentage: Option<u8>,
+    fee_percentage: Option<u16>,
     fee_recipient: Option<String>,
 ) -> Result<Response, ContractError> {
     let mut config = CONFIG.load(deps.storage)?;
